refactor: one save primitive — campaign_save_scope moves to utils, add-dataset adopts it
Post-merge reconciliation of the #115 hardening with this branch's helper move:
campaign_save_scope (+ PendingSave, describe_result) now lives in utils beside its
sibling datalad_save_scope; the bare utils.datalad_save is gone (its --to-git-less
successor is the scope), and add-dataset commits its statefile write through the
scope — flock first, clean-in second, one attributable node out. Also removes the
merge's latent NameError (campaign_init redefined run without subprocess imported).
